### 2024-06-28 #6942 @editorialbot
[PRE REVIEW]: Adamantine 1.0: A Thermomechanical Simulator for Additive Manufacturing
**Submitting author:** @Rombur (<a href="http://orcid.org/0000-0001-5954-6313">Bruno Turcksin</a>)
**Repository:** https://github.com/adamantine-sim/adamantine
**Branch with paper.md** (empty if default branch): release/1.0
**Version:** 1.0
