### 2024-07-11 a01cbc5
Store ScanPath on host
