+++ tests/test_material_deposition.cc
@@ -352,12 +352,11 @@ BOOST_AUTO_TEST_CASE(material_deposition)
 }
 
 std::pair<Kokkos::View<adamantine::ScanPathSegment *, Kokkos::HostSpace>,
-          adamantine::ScanPath<dealii::MemorySpace::Host>>
+          adamantine::ScanPath>
 create_scan_path(std::string file_name)
 {
   std::vector<adamantine::ScanPathSegment> scan_path_segments =
-      adamantine::ScanPath<dealii::MemorySpace::Host>::read_file(file_name,
-                                                                 "segment");
+      adamantine::ScanPath::read_file(file_name, "segment");
   Kokkos::View<adamantine::ScanPathSegment *, Kokkos::HostSpace>
       scan_path_segments_view("scan_path_segments", scan_path_segments.size());
   Kokkos::deep_copy(
