+++ source/GoldakHeatSource.cc
@@ -14,16 +14,15 @@
 namespace adamantine
 {
 
-template <int dim, typename MemorySpaceType>
-GoldakHeatSource<dim, MemorySpaceType>::GoldakHeatSource(
-    BeamHeatSourceProperties const &beam,
-    ScanPath<MemorySpaceType> const &scan_path)
+template <int dim>
+GoldakHeatSource<dim>::GoldakHeatSource(BeamHeatSourceProperties const &beam,
+                                        ScanPath const &scan_path)
     : _beam(beam), _scan_path(scan_path)
 {
 }
 
-template <int dim, typename MemorySpaceType>
-void GoldakHeatSource<dim, MemorySpaceType>::update_time(double time)
+template <int dim>
+void GoldakHeatSource<dim>::update_time(double time)
 {
   static const double pi_over_3_to_1p5 =
       std::pow(dealii::numbers::PI / 3.0, 1.5);
@@ -35,10 +34,9 @@ void GoldakHeatSource<dim, MemorySpaceType>::update_time(double time)
            (this->_beam.radius_squared * this->_beam.depth * pi_over_3_to_1p5);
 }
 
-template <int dim, typename MemorySpaceType>
-double
-GoldakHeatSource<dim, MemorySpaceType>::value(dealii::Point<dim> const &point,
-                                              double const height) const
+template <int dim>
+double GoldakHeatSource<dim>::value(dealii::Point<dim> const &point,
+                                    double const height) const
 {
   double const z = point[axis<dim>::z] - height;
   if ((z + this->_beam.depth) < 0.)
@@ -65,5 +63,4 @@ GoldakHeatSource<dim, MemorySpaceType>::value(dealii::Point<dim> const &point,
 }
 } // namespace adamantine
 
-INSTANTIATE_DIM_DEVICE(GoldakHeatSource)
-INSTANTIATE_DIM_HOST(GoldakHeatSource)
+INSTANTIATE_DIM(GoldakHeatSource)
