+++ source/ElectronBeamHeatSource.cc
@@ -16,8 +16,10 @@ namespace adamantine
 
 template <int dim>
 ElectronBeamHeatSource<dim>::ElectronBeamHeatSource(
-    BeamHeatSourceProperties const &beam, ScanPath const &scan_path)
-    : _beam(beam), _scan_path(scan_path)
+    boost::property_tree::ptree const &database)
+    : _beam(database),
+      _scan_path(database.get<std::string>("scan_path_file"),
+                 database.get<std::string>("scan_path_file_format"))
 {
 }
 
