+++ tests/test_material_deposition.cc
@@ -351,24 +351,9 @@ BOOST_AUTO_TEST_CASE(material_deposition)
   }
 }
 
-std::pair<Kokkos::View<adamantine::ScanPathSegment *, Kokkos::HostSpace>,
-          adamantine::ScanPath>
-create_scan_path(std::string file_name)
-{
-  std::vector<adamantine::ScanPathSegment> scan_path_segments =
-      adamantine::ScanPath::read_file(file_name, "segment");
-  Kokkos::View<adamantine::ScanPathSegment *, Kokkos::HostSpace>
-      scan_path_segments_view("scan_path_segments", scan_path_segments.size());
-  Kokkos::deep_copy(
-      scan_path_segments_view,
-      Kokkos::View<adamantine::ScanPathSegment *, Kokkos::HostSpace>{
-          scan_path_segments.data(), scan_path_segments.size()});
-  return {scan_path_segments_view, {scan_path_segments_view}};
-}
-
 BOOST_AUTO_TEST_CASE(deposition_from_scan_path_2d, *utf::tolerance(1e-13))
 {
-  auto [scan_path_segments, scan_path] = create_scan_path("scan_path.txt");
+  adamantine::ScanPath scan_path("scan_path.txt", "segment");
 
   boost::property_tree::ptree database;
   database.put("deposition_length", 0.0005);
@@ -416,7 +401,7 @@ BOOST_AUTO_TEST_CASE(deposition_from_scan_path_2d, *utf::tolerance(1e-13))
 
 BOOST_AUTO_TEST_CASE(deposition_from_scan_path_3d, *utf::tolerance(1e-13))
 {
-  auto [scan_path_segments, scan_path] = create_scan_path("scan_path.txt");
+  adamantine::ScanPath scan_path("scan_path.txt", "segment");
 
   boost::property_tree::ptree database;
   database.put("deposition_length", 0.0005);
@@ -468,7 +453,7 @@ BOOST_AUTO_TEST_CASE(deposition_from_scan_path_3d, *utf::tolerance(1e-13))
 
 BOOST_AUTO_TEST_CASE(deposition_from_L_scan_path_3d, *utf::tolerance(1e-13))
 {
-  auto [scan_path_segments, scan_path] = create_scan_path("scan_path_L.txt");
+  adamantine::ScanPath scan_path("scan_path_L.txt", "segment");
 
   boost::property_tree::ptree database;
   database.put("deposition_length", 0.0005);
@@ -554,8 +539,7 @@ BOOST_AUTO_TEST_CASE(deposition_from_L_scan_path_3d, *utf::tolerance(1e-13))
 BOOST_AUTO_TEST_CASE(deposition_from_diagonal_scan_path_3d,
                      *utf::tolerance(1e-10))
 {
-  auto [scan_path_segments, scan_path] =
-      create_scan_path("scan_path_diagonal.txt");
+  adamantine::ScanPath scan_path("scan_path_diagonal.txt", "segment");
 
   boost::property_tree::ptree database;
   database.put("deposition_length", 0.0005);
