+++ tests/test_scan_path.cc
@@ -21,11 +21,13 @@ class ScanPathTester
 public:
   std::vector<ScanPathSegment> get_segment_format_list()
   {
-    return ScanPath::read_file("scan_path.txt", "segment");
+    ScanPath scan_path("scan_path.txt", "segment");
+    return scan_path._segment_list;
   };
   std::vector<ScanPathSegment> get_event_series_format_list()
   {
-    return ScanPath::read_file("scan_path_event_series.inp", "event_series");
+    ScanPath scan_path("scan_path_event_series.inp", "event_series");
+    return scan_path._segment_list;
   };
 };
 
@@ -73,11 +75,7 @@ BOOST_AUTO_TEST_CASE(scan_path, *utf::tolerance(1e-12))
 
 BOOST_AUTO_TEST_CASE(scan_path_location, *utf::tolerance(1e-10))
 {
-  std::vector<ScanPathSegment> scan_path_segments =
-      ScanPath::read_file("scan_path.txt", "segment");
-  Kokkos::View<ScanPathSegment *, Kokkos::HostSpace> scan_paths_segments_view(
-      scan_path_segments.data(), scan_path_segments.size());
-  ScanPath scan_path(scan_paths_segments_view);
+  ScanPath scan_path("scan_path.txt", "segment");
   double time = 1.0e-7;
   dealii::Point<3> p1 = scan_path.value(time);
 

[…]
