### 2024-07-18 #6942 @lucydot
@nrichart + @lukashof + @FelixFroelich + @ericneiva - Hello 👋 . Are you available to review this submission to JOSS?
