### 2024-07-22 #7017 @editorialbot
[REVIEW]: Adamantine 1.0: A Thermomechanical Simulator for Additive Manufacturing
**Submitting author:** @Rombur (<a href="http://orcid.org/0000-0001-5954-6313">Bruno Turcksin</a>)
**Repository:** https://github.com/adamantine-sim/adamantine
**Branch with paper.md** (empty if default branch): release/1.0
**Version:** v1.0

@mohd-afeef-badri & @ericneiva & @FelixFroelich & @lukashof, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
