### 2024-07-22 59bbbb5
Add documentation about input file formats
