### 2024-07-22 56a9718
Add documentation for element activation
