+++ algorithmic_choice.md
@@ -20,8 +20,19 @@ The implementation is done matrix-free for the following reasons:
 
 ## Adaptive mesh refinement
 Usually, the powder layer is about 50 microns thick but the piece that is being
-built is several centimeters long. Moreover, since the material is melted using
+built is tens of centimeters long. Moreover, since the material is melted using
 an electron beam or a laser, the melting zone is very localized. This means that
 a uniform mesh would require a very large number of cells in places where nothing
 happens (material not heated yet or already cooled). Using AMR, we can refine
 the zones that are of interest for a given point in time.
+
+## Element activation
+To simulate the addition on material, we use the hp-capability of 
+[deal.II](https://www.dealii.org). Deal.II supports a special kind of finite 
+element called `FE_Nothing`. `FE_Nothing` is a finite element that does with
+zero degree of freedom. This can be used to represent empty cells in a mesh on
+which no degrees of freedom should be allocated. To simulate the addition of
+material, we replace the `FE_Nothing` associated with a cell with a regular finite 
+element, i.e., we activate an element. Using this technique, the addition of 
+material can be done cell-wise. By coupling element activation and AMR, we can
+add arbitrary small amount of material.
