### 2024-07-22 efb1bb8
Split mechanical_physics->setup_dofs
