+++ tests/test_mechanical_physics.cc
@@ -481,8 +481,9 @@ run_eshelby(std::vector<dealii::Point<dim>> pts, unsigned int refinement_cycles)
 
   std::vector<bool> has_melted(triangulation.n_active_cells(), false);
 
-  mechanical_physics.setup_dofs(thermal_physics.get_dof_handler(), temperature,
-                                has_melted);
+  mechanical_physics.prepare_transfer(thermal_physics.get_dof_handler(),
+                                      temperature, has_melted);
+  mechanical_physics.setup_dofs();
 
   auto solution = mechanical_physics.solve();
 
