### 2024-07-22 773c53d
Use vector with locally relevant dofs in compute_stress
