### 2024-07-22 f2d286c
Use vector with locally relevant dofs in compute_stress
