### 2024-08-06 #7017 @lukashof
Hi, @Rombur,

I finally got around to taking a look at the software. I was using the provided container via Docker Desktop on a Windows machine. I tried to run a simulation using the provided `input.info`. The run completed almost immediately. Unfortunately, there wasn't a lot to see when examining the output with ParaView: Just a 2D rectangle divided in two subdomains. None of the fields seem to change. I might have overlooked something, but looking at the definition of the provided example, I assume this was provided merely to test whether the software runs?

I recognize that there are some instructions in the `README` to help to create a new model definition. For an outsider and deal.II-novice like myself, it would nevertheless be invaluable to have access to a few, more elaborate examples that showcase the intended use and capabilities of the software. That is especially true for some of the more advanced capabilities you mention in your paper, such as the data assimilation feature. That feature in particular would also require some measurement data to use, if I am not mistaken. Could you provide us with a few examples that showcase the different AM processes and capabilities of the software?

After (superficially) reading your submitted paper, I also believe that it would benefit if you showcased the use of the software a bit. I understand that it is meant as a brief description of the software, but some exemplary results and visualization would make it far easier to understand its capabilities.
