### 2024-08-06 #7017 @Rombur
@lukashof you are correct, the installed input file is there to check that the code runs nothing more. There are more examples of input files in [tests/data](https://github.com/adamantine-sim/adamantine/tree/release/1.0/tests/data). We should have documented it somewhere.
The `demo_316_short.info` input file which uses `demo_316_short_scan_path.txt` is a more interesting case. The `bare_plate_L_da.info` input file uses data assimilation. To run the code you need need several other `bare_place_L` files but they are all present in [tests/data](https://github.com/adamantine-sim/adamantine/tree/release/1.0/tests/data)
