### 2024-08-14 #7017 @lukashof
Hi @Rombur,

I now ran the examples you mentioned.

This is a minor detail, but I am not sure how to best visualize the results. I have no issues opening the .pvd-files in ParaView. So far, I found that `Surface` or `Surface with Edges` representations with the opacity mapping set to hide the non-activated cells works best. But the result looks a bit “glitchy” when looking at it from the wrong side. Is there some setting I am missing to avoid this?
![grafik](https://github.com/user-attachments/assets/b56657bd-2887-474c-aa65-7d569e663053)

I also tried to run the `thermoelastic_bare_plate` example, as it seems to include a coupled thermomechanical simulation. Unfortunately, I got an [error message](https://github.com/user-attachments/files/16613803/error_thermoelastic_bare_plate.md) that is a bit cryptic to me.

These examples are obviously far more helpful than the simple test. But I have to say that as a new user of the software, I find it hard to understand and gauge the capabilities of the tool, as they are presented now. I believe the project would be more accessible if you provided some explanations as to what each example is supposed to demonstrate.

From what I saw so far, these examples are also too simplistic to demonstrate the capabilities you claim in your paper. Running the `demo_316_short` example for instance, didn't help me understand how the data assimilation works or demonstrate how it improves the results compared to a model without it.

@FelixFroelich and I will be out of office and unavailable for the coming two weeks.
