### 2024-08-19 #7017 @Rombur
@lukashof

> This is a minor detail, but I am not sure how to best visualize the results. I have no issues opening the .pvd-files in ParaView. So far, I found that Surface or Surface with Edges representations with the opacity mapping set to hide the non-activated cells works best. But the result looks a bit “glitchy” when looking at it from the wrong side. Is there some setting I am missing to avoid this?

I don't use Paraview, I use Visit instead. In Visit, you cannot see which cells are active but you can mask the cells outside of a threshold. For the non-active cells, the temperature is set to zero, so I just hide the cells that have a temperature below one. I assume that Paraview has a similar capability.

> I also tried to run the thermoelastic_bare_plate example, as it seems to include a coupled thermomechanical simulation. Unfortunately, I got an [error message](https://github.com/user-attachments/files/16613803/error_thermoelastic_bare_plate.md) that is a bit cryptic to me.

I see that you are running the test with two processors but currently, we do not support MPI for (thermo)mechanical simulation. This is explained in the `MPI support` section of the paper:
```
While mechanical and thermomechanical simulations are limited to serial execution, thermal and EnKF ensemble simulations can use MPI.
```

> These examples are obviously far more helpful than the simple test. But I have to say that as a new user of the software, I find it hard to understand and gauge the capabilities of the tool, as they are presented now. I believe the project would be more accessible if you provided some explanations as to what each example is supposed to demonstrate.

I'll add more documentation on the website.

> From what I saw so far, these examples are also too simplistic to demonstrate the capabilities you claim in your paper. Running the demo_316_short example for instance, didn't help me understand how the data assimilation works or demonstrate how it improves the results compared to a model without it.

I am not sure how to satisfy your expectations. The paper must be two pages or less, there is no room to explain the theory behind data assimilation. We have implemented a well-known method of data assimilation and provided a source.   The `demo_316_short` example does not use data assimilation. Other examples such as the `bare_plate_L_da.info` input file that I referenced earlier in this issue do use data assimilation.
