+++ input_file.md
@@ -1,0 +1,169 @@
+---
+layout: default
+parent: Run
+title: Input file
+nav_order: 1
+usemathjax: true
+---
+
+# Input file
+*adamantine* supports Boost INFO format and json. The input file is assumed to use the INFO format unless the file extension is `.json`.
+The following options are available:
+* **boundary** (required):
+  * **type**: type of boundary: `adiabatic`, `radiative`, or `convective`. Multiple types
+  can be chosen simultaneously by separating them by comma (required)
+* **physics** (required):
+  * **thermal**: thermal simulation: true or false (required)
+  * **mechanical**: mechanical simulation: true or fase (required)
+  * if both thermal and mechanical parameters are true, solve a coupled thermo-mechanics problem
+* **discretization** (required):
+  * **thermal** (required if *physics.thermal* is true):
+    * **fe\_degree**: degree of the finite element used (required)
+    * **quadrature**: quadrature used: `gauss` or `lobatto` (default value: `gauss`)
+  * **mechanical** (required if *physics.mechanical* is true):
+    * **fe\_degree**: degree of the finite element used (required)
+* **geometry** (required):
+  * **dim**: the dimension of the problem (2 or 3, required)
+  * **material\_height**: below this height the domain contains material. Above this
+  height the domain is empty (default value: 1e9)
+  * **use\_powder**: the additive manufacturing process use powder: true or false
+  (default value: false)
+  * if use\_powder is true:
+    * **powder\_layer**: thickness of the initial layer of powder in meters (required)
+  * **material\_deposition**: material is deposed during the simulation: true or
+  false (default value: false)
+  * if material\_deposition is true:
+    * **material\_deposition\_method**: `file` or `scan_paths`
+    * if material\_deposition\_method is file:
+        * **material\_deposition\_file**: material deposition filename
+    * if material\_deposition\_method is scan\_paths:
+        * **deposition\_length**: length of material deposition boxes along the scan direction (required)
+        * **deposition\_width**: width of material deposition boxes (in the plane of the material, normal to the scan direction, 3D only) (required)
+        * **deposition\_height**: height of material deposition boxes (out of the plane of the material) (required)
+        * **deposition\_lead\_time**: amount of time before the scan path reaches a point that the material is added (required)
+        * **deposition\_time**: add the material in bigger lumps (optional)
+  * **import\_mesh**: true or false (required)
+  * if import\_mesh is true:
+    * **mesh\_file**: The filename for the mesh file (required)
+    * **mesh\_format**: `abaqus`, `assimp`, `unv`, `ucd`, `dbmesh`, `gmsh`, `tecplot`, `xda`, `vtk`,
+    `vtu`, `exodus`, or `default`, i.e., use the file suffix to try to determine the
+    mesh format (required)
+    * **mesh\_scale\_factor**: Apply a uniform scaling factor to the mesh (e.g. if the mesh is defined
+    in mm or inches instead of m) (default value: 1)
+    * **reset\_material\_id**: Clear the material IDs defined in the mesh and set them all to zero so 
+      all material properties are given by the `material_0` input block: true or false (default value: false)
+  * if import\_mesh is false:
+    * **length**: the length of the domain in meters (required)
+    * **height**: the height of the domain in meters (required)
+    * **width**: the width of the domain in meters (only in 3D)
+    * **length\_origin**: the reference location in the length direction (default value: 0)
+    * **height\_origin**: the reference location in the height direction (default value: 0)
+    * **width\_origini**: the reference location in the width direction (only in 3D) (default value: 0)
[…]
