### 2024-08-20 #7017 @lucydot
> @FelixFroelich and I will be out of office and unavailable for the coming two weeks.

Thanks @lukashof and @FelixFroelich for review so far and the update.

> The paper must be two pages or less, there is no room to explain the theory behind data assimilation. We have implemented a well-known method of data assimilation and provided a source. The demo_316_short example does not use data assimilation. Other examples such as the bare_plate_L_da.info input file that I referenced earlier in this issue do use data assimilation.

@Rombur you are correct that the paper is the place for the underlying theory.
Looking at the online documentation and repo, I can see why it may be difficult for a potential user to connect the tutorials with the wider documentation. You could, for example, link from the "Data assimilation" page on your jekyll site to the tutorials which demonstrate data assimilation. Usually we expect a tutorial page which outlines/summarises the tutorials and the order in which they might be approached, or what particular features they demonstrate. It indicates to users where best to start so they are not searching through several files.
