### 2024-08-21 #7017 @mohd-afeef-badri
Dear @Rombur,

I have finally had time since couple of weeks now to review your paper and repository.

Firstly, I would like to commend you and your coauthor for this work and a well written JOSS paper. Although concise, the paper provides insightful details to try your simulation code. After using the code myself for a while now, I am convinced that it holds significant potential for contributing to other research and engineering projects.

Despite the paper's and your codes strengths, there are some minor aspects that I believe need  improvement before I can recommend this for publication. I have outlined a few concerns that I  hope you will address in the revised version of the paper and within your code:

## Minor concerns in the article ##
#### Clarification and Additional Content ####
1. For readers interested in the mathematical aspects of the thermal, mechanical, and thermomechanical simulations, it might be helpful to cite your [jekyll site](https://adamantine-sim.github.io/adamantine/doc/governing_equations.html) that delves into these details.

2. Including a brief section that outlines the types of meshes supported by Adamantine and what users can expect during post-processing would significantly enhance the paper's utility for potential users. Perhaps and example simulations that demonstrates this.

3. The statement that "part of the thermal simulation can be performed on the GPU" could benefit from further elaboration. Specifically, detailing which aspects of the thermal simulation are GPU-accelerated and under what conditions this occurs would clarify this feature for users. Since Thermal simulations seem to be the one supporting the CPU and GPU parallelism, does it also support MPI-GPU mixed parallelism ?

#### References ####
1. The URLs provided for "Keller et al." and "Knapp et al." appear to be incorrect. Please review and update them.
2. The reference for "Asch et al." lists an incorrect publisher. Please update this reference with the correct publisher.
3. The reference for "Trilinos" is missing the publication date. Please include this information.
4. You’ve referenced Abaqus but not Ansys. It may be more balanced to either reference both or omit both, depending on the context.

#### Text corrections ####
1. There is a minor typo in the "Thermomechanical simulation" section on line 62: "fosr" should be corrected to "for."
2. The acronyms "IR" and "AM" are introduced but not subsequently used in the text. It might be clearer to avoid introducing these acronyms if they are not utilized later. Additionally, the acronym for "Ensemble Kalman Filter" (EnKF) is introduced multiple times. Consider introducing it only once for clarity.

## Concerns While Testing the Code ##
1. MPI and GPU Testing Issues: While the paper claims support for MPI and GPU acceleration in thermal simulations, I was unable to verify this using the "HourGlass_AOP" example, which I believe is a thermal simulation. Could you clarify if this example should demonstrate MPI and GPU acceleration, or if there are specific conditions or configurations required to observe these features? Is it too small simulation to observe the parallelization feature ?

- I did not observe any GPU activity when running the "HourGlass_AOP" example (monitored via nvidia-smi and nvtop). Is this due to the test case being too small, or is GPU usage not supported within the provided Docker image?

- My initial scaling tests did not show significant improvements with increased processors. Could this be due to the small size of the test case, or might there be other factors affecting scaling in this particular example? Clarifying this in the documentation could help users better understand the performance expectations.

      | N_proc | Time Total |
      | ------ | ---------- |
      | 1      | 3m42.377s  |
      | 2      | 3m48.103s  |
      | 4      | 3m27.661s  |
      | 8      | 3m51.990s  |

2.  For users who are less familiar with ParaView, it might be useful to add a brief note explaining that the '.pvd' file should be opened in ParaView, as it contains the relevant fields "A, B, C." This will assist users in navigating the '.vtu,' '.pvtu,' and '.pvd' files, which may otherwise be confusing for novices.

3. I second suggestions from @lukashof  and @lucydot, that the inclusion of a small tutorial section to guide novice users on how to get started with the code would be highly valuable. Additionally, a README table describing the tests in the `tests/data` directory, categorizing them by type (e.g., parallel/serial, DA/non-DA, type of simulation), would help users quickly understand the purpose of each test without needing to inspect individual files.

4. While the provision of a Docker image is appreciated, a brief explanation on how to run the code using this Docker image would be beneficial for users, particularly those unfamiliar with Docker.

I look forward to seeing the updates /responded from you and hope my feedback contributes to further refining this work.
