### 2024-08-22 f907bca
Add HourGlass example
