+++ doc/examples.md
@@ -1,0 +1,17 @@
+---
+layout: page
+title: Examples
+nav_order: 4
+has_children: true
+---
+
+# Examples
+By default in the *adamantine* install directory, you should find the `input.info`
+input file. This input file can be used to quickly check that *adamantine* runs but 
+the simulation is not very exciting. It is a purely thermal simulation on a 2D
+domain.
+
+More examples can be found in
+[tests/data](https://github.com/adamantine-sim/adamantine/tree/master/tests/data).
+We showcase a few of them below.
+
