### 2024-08-23 2916e8f
Remove wrong sentence
