### 2024-08-23 adamantine-sim/adamantine#306 @Rombur
JOSS review todo list
Link to the review https://github.com/openjournals/joss-reviews/issues/7017

- Website:
	- [x] Add documentation on how to visualize the solution
	- [x] Add documentation on the examples found in `test/data`
	- [x] Rework the `HourGlass` test with updated data
	- [x] Add more documentation on data assimilation and link to examples
	- [x] Add explanation on how to use Docker and the issue with creating Docker container for GPU
	- [x] In the MPI section, explain the partitioning better
- Paper:
	- [x] Add link to the website in the paper
	- [x] Add section about meshes and postprocessing in the paper
	- [x] Fix references
	- [x] Fix typos
