### 2024-08-23 #7017 @Rombur
@mohd-afeef-badri Thanks for your comments. I have open an issue [here](https://github.com/adamantine-sim/adamantine/issues/306) to keep track of what needs to be done.

It is normal that you do not see any GPU activity when using the Docker image. I forgot to document it but the Docker image cannot run on GPU. The issue is that we cannot create a fat binary that works for all GPUs. Instead, we have to create a Docker image for each architecture (volta, ampere, etc.) . On top of that if you have an NVidia GPU, you also need to install NVidia's Docker plugin. This is not very difficult to do but we just thought that it wasn't worth the effort to provide Docker images with GPU support. You do not have these issue if you build the code yourself though that it is a lot more work. If you tell me which GPU you are using, I can create a Docker image with GPU support for you.

I will rework the HourGlass test to show scaling. The simulation is probably not running for long enough and so, the domain is too small. We distribute the active cells through the available processors. At the beginning of the build, there are very few cells to distributes and you pay for the overhead. Once you have build more of the object, you have more cells to distribute and this is not an issue anymore. We used to distributed the cells regardless of their active/inactive status but then, some processors would only own inactive cells and were idle for the most of  the build.
