### 2024-08-26 #7017 @lucydot
@mohd-afeef-badri, @ericneiva, @FelixFroelich, @lukashof - is one of you able to build code (not necessarily for GPU) to confirm that guidance is sufficient? would like at least one attempt at building (rather than using Docker image) as part of the review process.
