### 2024-08-26 75f2b29
Add missing installation requirements
