### 2024-08-26 #7017 @Rombur
@lucydot Do you want them to build only our code or the entire software stack? I have docker images with all the dependencies already compiled but without our code. We have `rombur/adamantine-stack:latest` with the stack compiled to use A100 and `rombur/adamantine-stack:no_gpu-latest` which is the stack that we use to create the Docker image of adamantine.
