+++ doc/examples.md
@@ -1,7 +1,7 @@
 ---
 layout: page
 title: Examples
-nav_order: 4
+nav_order: 5
 has_children: true
 ---
 
