### 2024-08-29 3f67ed2
Fix favicon
