+++ _config.yml
@@ -38,7 +38,6 @@ compress_html:
 
 
 logo: "https://raw.githubusercontent.com/adamantine-sim/adamantine/master/doc/logo/adamantine_wordmark.png"
-favicon_ico: "https://raw.githubusercontent.com/adamantine-sim/adamantine/master/doc/logo/adamantine_logo.png"
 
 # Aux links for the upper right navigation
 aux_links:
