### 2024-08-29 386ba28
Adds adamantine executable to container image
