### 2024-08-30 #7017 @mohd-afeef-badri
> @mohd-afeef-badri Thanks for your comments. I have open an issue [here](https://github.com/adamantine-sim/adamantine/issues/306) to keep track of what needs to be done.

@Rombur, thank you for creating the [issue](https://github.com/adamantine-sim/adamantine/issues/306). This will definitely be useful for validating the review process later on.

> It is normal that you do not see any GPU activity when using the Docker image. I forgot to document it but the Docker image cannot run on GPU. The issue is that we cannot create a fat binary that works for all GPUs. Instead, we have to create a Docker image for each architecture (volta, ampere, etc.) . On top of that if you have an NVidia GPU, you also need to install NVidia's Docker plugin. This is not very difficult to do but we just thought that it wasn't worth the effort to provide Docker images with GPU support. You do not have these issue if you build the code yourself though that it is a lot more work. If you tell me which GPU you are using, I can create a Docker image with GPU support for you.

Got it; that explains the lack of GPU activity with the current Docker image. If you could provide a Docker image with NVIDIA GPU support, it would be highly appreciated, but please feel free to decline if it’s too much trouble.

In the meantime, I was able to compile adamantine using CMake. The whole compilation process was a  bit tricky and time-taking, but I must admit the Dockerfile script (`ci/Dockerfile`) to get all the right options needed for full compilation of the dependencies helped.   Concerning this, please note that the HDF5 URL is broken (`ci/Dockerfile`), so I encounter issues due to missing sources from the provided link I had to adapt this please correct this in your repository. Also I had to adapt according to my version of GPU for Trilinos compilation.

Since now I have a locally compiled version of the code, I’ll reassess your adamantine tests with it. This I shall do once you have made your changes in the [issue](https://github.com/adamantine-sim/adamantine/issues/306).

> I will rework the HourGlass test to show scaling. The simulation is probably not running for long enough and so, the domain is too small. We distribute the active cells through the available processors. At the beginning of the build, there are very few cells to distributes and you pay for the overhead. Once you have build more of the object, you have more cells to distribute and this is not an issue anymore. We used to distributed the cells regardless of their active/inactive status but then, some processors would only own inactive cells and were idle for the most of the build.

I understand your point. I look forward to seeing your changes and reassessing the scaling performance.

> @mohd-afeef-badri, @ericneiva, @FelixFroelich, @lukashof - is one of you able to build code (not necessarily for GPU) to confirm that guidance is sufficient? would like at least one attempt at building (rather than using Docker image) as part of the review process.

@lucydot, I was able to build adamantine using the right options from the Dockerfile script (`ci/Dockerfile`). However, I feel this process is quite challenging due to the need for a specifically compiled deal-ii with P4EST, Arborx, and Trilinos support. Also, it took few hours for the compilation to finish, I do not know if that is normal. This complexity is why the Docker solution here is a more practical option.

@Rombur, I recommend providing a sample compile script for users to reference. Converting the Dockerfile (`ci/Dockerfile`) into a bash script and sharing it as an example could greatly assist users in setting up their builds with the correct options. Also a warning on long compile time should be provided.
