### 2024-09-02 #7017 @lucydot
@Rombur; thanks. @mohd-afeef-badri has done wonderful work building adamantine and giving feedback. The suggestion re: providing a sample compile script seems sensible. Although most users are likely to go down the Docker route, we want to ensure compilation is possible so that the software, as open source, can also be easily adapted/extended by others.
