### 2024-09-03 #7017 @Rombur
> If you could provide a Docker image with NVIDIA GPU support, it would be highly appreciated, but please feel free to decline if it’s too much trouble.

I can create a new image but I need to know if it's for a V100, a A100, etc. I cannot create one image that works for all Nvidia GPUs.

@lucydot I am quite reluctant to provide a script to compile my code and all its dependencies. It never works for everybody but that's exactly what people will (rightfully) expect it to do. That's what one of my dependencies did. They started a small bash script to help people to compile the code and its dependencies. Now it's thousands of lines of code and patches and it still does not work all the time. My code is pretty straightforward to compile, it's installing all the dependencies that's difficult.
