### 2024-09-05 #7017 @lucydot
@Rombur ; your point here seems reasonable. It is good that @mohd-afeef-badri could compile, as it indicates that you have documented the dependencies required sufficiently, even if dependency installation itself is still a little painful.
