### 2024-09-09 031958d
Add documentation about GPU computation
