### 2024-09-16 #7017 @lucydot
@Rombur  - I can see you are working through a [TODO list](https://github.com/adamantine-sim/adamantine/issues/306) relating to the paper and documentation.

Whilst you are doing this, and now installation seems to be largely addressed, @mohd-afeef-badri, @ericneiva, @FelixFroelich, @lukashof how are your reviews going? In particular, does the software function and perform as described? There has been some discussion of this further up the thread, but I also see some unticked boxes (and thank you for your efforts so far!)
