### 2024-09-19 #7017 @mohd-afeef-badri
> I can create a new image but I need to know if it's for a V100, a A100, etc. I cannot create one image that works for all Nvidia GPUs.

@Rombur Thanks for the response, a A100 image would be great to test your tools on GPU.

@lucydot From my side all is ok, I am waiting for the TODO list f to be finished, so that I can retest everything and this is when you can expect my check-list to be completed.
