+++ source/MechanicalPhysics.cc
@@ -44,22 +44,10 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
   // Solve the mechanical problem only on the part of the domain that has solid
   // material.
-  unsigned int n_local_cells = 0;
-  for (auto const &cell :
-       dealii::filter_iterators(_dof_handler.active_cell_iterators(),
-                                dealii::IteratorFilters::LocallyOwnedCell()))
-  {
-    if (_material_properties.get_state_ratio(
-            cell, MaterialStates::State::solid) > 0.99)
-    {
-      cell->set_active_fe_index(0);
-      ++n_local_cells;
-    }
-    else
-    {
-      cell->set_active_fe_index(1);
-    }
-  }
+
+  std::cout << "Setting active fe_indices mechanical 0" << std::endl;
+
+  unsigned int n_local_cells = set_active_fe_indices();
 
   // Create the mechanical operator
   _mechanical_operator = std::make_unique<
@@ -87,8 +75,24 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
 void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
+    setup_dofs(
+        dealii::DoFHandler<dim> const &thermal_dof_handler,
+        dealii::LA::distributed::Vector<double, dealii::MemorySpace::Host> const
+            &temperature,
+        std::vector<bool> const &has_melted,
+        std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
+{
+  _mechanical_operator->update_temperature(thermal_dof_handler, temperature,
+                                           has_melted);
+  setup_dofs(body_forces);
+}
+
+template <int dim, int p_order, typename MaterialStates,
+          typename MemorySpaceType>
+void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
     setup_dofs(std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
 {
+  std::cout << "Distribute dofs mechanical" << std::endl;
   _dof_handler.distribute_dofs(_fe_collection);
   dealii::IndexSet locally_relevant_dofs;
   dealii::DoFTools::extract_locally_relevant_dofs(_dof_handler,
@@ -112,16 +116,73 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
-void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    setup_dofs(
-        dealii::DoFHandler<dim> const &thermal_dof_handler,
-        dealii::LA::distributed::Vector<double, dealii::MemorySpace::Host> const
-            &temperature,
-        std::vector<bool> const &has_melted,
-        std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
+unsigned int MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
+    set_active_fe_indices(
+        std::optional<std::reference_wrapper<dealii::DoFHandler<dim> const>>
+            thermal_dof_handler)
+
 {
-  _mechanical_operator->update_temperature(thermal_dof_handler, temperature,
-                                           has_melted);
+  unsigned int n_local_cells = 0;
+  for (auto const &cell :
+       dealii::filter_iterators(_dof_handler.active_cell_iterators(),
+                                dealii::IteratorFilters::LocallyOwnedCell()))
+  {
+    auto current_fe_index = cell->future_fe_index();
+    if (_material_properties.get_state_ratio(
[…]
