### 2024-09-19 7f899bc
Improve HourGlass documentation
