+++ doc/examples/hour_glass.md
@@ -2,7 +2,7 @@
 layout: default
 parent: Examples
 title: HourGlass
-nav_order: 1
+nav_order: 2
 ---
 
 <head>
@@ -28,10 +28,15 @@ nav_order: 1
 
 
 # HourGlass
+This example is similar to the
+[Demo316]({{site.baseurl}}/doc/examples/demo_316) example. This is also a purely
+thermal problem. We will use this example to understand which parameters affects
+the performance of *adamantine*.
+
 This example is composed of the following files:
- * **HourGlass\_AOP.info:** the input file
- * **HourGlass\_AOP.vtk:** the mesh
- * **HourGlass\_AOP\_scan\_path.txt:** the scan path of the heat source
+ * **HourGlass_AOP.info:** the input file
+ * **HourGlass_AOP.vtk:** the mesh
+ * **HourGlass_AOP_scan_path.txt:** the scan path of the heat source
 
 The domain is an hour glass with an hole in the center.
 
@@ -52,6 +57,165 @@ The domain is an hour glass with an hole in the center.
  </div>
 </div> 
 
-This example is a purely thermal problem. By changing the
-`time_stepping.duration` parameter, you can choose how much of the manufacturing
-process is simulated.
+To understand where most of the time is spent, we will use [caliper](http://software.llnl.gov/Caliper).
+First, we run **HourGlass_AOP.info** using a single processor. *Caliper* returns the following results:
+
+|Path                                   | Min time/rank | Max time/rank | Avg time/rank | Time % | 
+|:--------------------------------------|:--------------|:--------------|:--------------|:-------|
+|main                                   |    106.35     |    106.35     |    106.35     | 100.00 | 
+|<pre> run</pre>                        |    106.35     |    106.35     |    106.35     |  99.99 | 
+|<pre>  output_pvtu</pre>               |      0.17     |      0.17     |      0.17     |   0.16 | 
+|<pre>  main_loop</pre>                 |    101.01     |    101.01     |    101.01     |  94.97 | 
+|<pre>   refine_mesh</pre>              |      0.06     |      0.06     |      0.06     |   0.06 | 
+|<pre>   add_material</pre>             |     78.80     |     78.80     |     78.80     |  74.09 | 
+|<pre>    refine triangulation</pre>    |     31.98     |     31.98     |     31.98     |  30.07 |
+|<pre>   evaluate_thermal_physics</pre> |      0.78     |      0.78     |      0.78     |   0.73 | 
+|<pre>   output_pvtu</pre>              |     16.57     |     16.57     |     16.57     |  15.58 | 
+
+We see that the simulation took 105 seconds of which 74% is spent in `add_material`.
+The core of computation is `evaluate_thermal_physics` which took less than 1% of
+the total simulation time. Writing the output files, `output_pvtu`, took another
+15%. Neither `add_material` or `output_pvtu` scale well with the number of processors.
+Using more processors will not improve the performance significantly. Moreover
+there are only 1241 degrees of freedom which is too small to take advantage of
+more processors. 
+
+Let's modified **HourGlass_AOP.info**:
+ - add `deposition_time 25` inside `geometry`. Using this option, the material 
+ is deposited in such a way that it will take 20 seconds for the heat source 
+ to reach the end of the deposition. This decreases the number of time, material
+ is added which will greatly speedup the simulation.
+ - change `time_steps_between_output` from 10 to 500.
+ - change `duration` to 500 s. By running the simulation for a longer time, more
+ cells will be activated and we will have more degrees of freedom to spread
+ between the processors.
+
+When using one processor, we get:
+
+|Path                                   | Min time/rank | Max time/rank | Avg time/rank | Time % | 
+|:--------------------------------------|:--------------|:--------------|:--------------|:-------|
[…]
