+++ doc/algorithmic_choice.md
@@ -49,9 +49,10 @@ The mesh is partitioned using [p4est](https://www.p4est.org/). The partitioning
 takes into account that there is nothing to do on inactive cells. It tries to
 distribute evenly the active cells between all the processors. If at the beginning
 of the simulation very few cells are active, each processor will own a very
-small number of cells. In this case, the communication cost can be important
+small number of active cells. In this case, the communication cost can be important
 compared to the computation cost. Once more cells are activated this ceases to
-be a problem.
+be a problem. For a more in-depth discussion take a look at the [HourGlass
+example]({{site.baseurl}}/doc/examples/hour_glass).
 
 ### EnKF
 For EnKF ensemble simulations, the partitioning scheme works as follows:
