### 2024-09-19 cfeaf44
Fix white spaces in the tables
