+++ doc/examples/hour_glass.md
@@ -60,17 +60,17 @@ The domain is an hour glass with an hole in the center.
 To understand where most of the time is spent, we will use [caliper](http://software.llnl.gov/Caliper).
 First, we run **HourGlass_AOP.info** using a single processor. *Caliper* returns the following results:
 
-|Path                                   | Min time/rank | Max time/rank | Avg time/rank | Time % | 
-|:--------------------------------------|:--------------|:--------------|:--------------|:-------|
-|main                                   |    106.35     |    106.35     |    106.35     | 100.00 | 
-|<pre> run</pre>                        |    106.35     |    106.35     |    106.35     |  99.99 | 
-|<pre>  output_pvtu</pre>               |      0.17     |      0.17     |      0.17     |   0.16 | 
-|<pre>  main_loop</pre>                 |    101.01     |    101.01     |    101.01     |  94.97 | 
-|<pre>   refine_mesh</pre>              |      0.06     |      0.06     |      0.06     |   0.06 | 
-|<pre>   add_material</pre>             |     78.80     |     78.80     |     78.80     |  74.09 | 
-|<pre>    refine triangulation</pre>    |     31.98     |     31.98     |     31.98     |  30.07 |
-|<pre>   evaluate_thermal_physics</pre> |      0.78     |      0.78     |      0.78     |   0.73 | 
-|<pre>   output_pvtu</pre>              |     16.57     |     16.57     |     16.57     |  15.58 | 
+|Path                                                                 | Min time/rank | Max time/rank | Avg time/rank | Time % | 
+|:--------------------------------------------------------------------|:-------------:|:-------------:|:-------------:|:------:|
+|main                                                                 |    106.35     |    106.35     |    106.35     | 100.00 | 
+|&nbsp;&nbsp;run                                                      |    106.35     |    106.35     |    106.35     |  99.99 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;output_pvtu                                  |      0.17     |      0.17     |      0.17     |   0.16 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;main_loop                                    |    101.01     |    101.01     |    101.01     |  94.97 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;refine_mesh                      |      0.06     |      0.06     |      0.06     |   0.06 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;add_material                     |     78.80     |     78.80     |     78.80     |  74.09 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;refine triangulation |     31.98     |     31.98     |     31.98     |  30.07 |
+|&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;evaluate_thermal_physics         |      0.78     |      0.78     |      0.78     |   0.73 | 
+|&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;output_pvtu                      |     16.57     |     16.57     |     16.57     |  15.58 | 
 
 We see that the simulation took 105 seconds of which 74% is spent in `add_material`.
 The core of computation is `evaluate_thermal_physics` which took less than 1% of
@@ -92,33 +92,33 @@ Let's modified **HourGlass_AOP.info**:
 
 When using one processor, we get:
 
-|Path                                   | Min time/rank | Max time/rank | Avg time/rank | Time % | 
-|:--------------------------------------|:--------------|:--------------|:--------------|:-------|
[…]
