+++ source/MechanicalPhysics.cc
@@ -170,7 +170,36 @@ void MechanicalPhysics<dim, p_order, MaterialStates,
                    _mechanical_operator->rhs().get_mpi_communicator());
 
   displacement.zero_out_ghost_values();
-  _solution_transfer.interpolate(displacement);
+
+  // Update _old_displacement if necessary
+  _old_displacement.reinit(_mechanical_operator->rhs().get_partitioner());
+  if (_saved_old_displacement.size())
+  {
+    unsigned int const n_dofs_per_cell = _fe_collection.max_dofs_per_cell();
+
+    std::vector<dealii::types::global_dof_index> global_dof_indices(
+        n_dofs_per_cell);
+
+    unsigned int cell_id = 0;
+    for (auto const &cell :
+         dealii::filter_iterators(_dof_handler.active_cell_iterators(),
+                                  dealii::IteratorFilters::LocallyOwnedCell()))
+    {
+      auto fe_index = cell->active_fe_index();
+      if (fe_index == 0)
+      {
+        cell->get_dof_indices(global_dof_indices);
+        for (unsigned int i = 0; i < n_dofs_per_cell; ++i)
+        {
+          _old_displacement[global_dof_indices[i]] =
+              _saved_old_displacement[cell_id][i];
+        }
+      }
+      ++cell_id;
+    }
+    _old_displacement.compress(dealii::VectorOperation::insert);
+  }
+
   displacement.update_ghost_values();
   std::swap(_old_displacement, displacement);
 }
@@ -186,7 +215,7 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
   // Update the active fe indices, the plastic variables, and the displacement.
   unsigned int const n_quad_pts = _q_collection.max_n_quadrature_points();
   unsigned int cell_id = 0;
-  std::vector<std::vector<double>> saved_old_displacement;
+  // std::vector<std::vector<double>> saved_old_displacement;
   std::vector<std::vector<double>> tmp_plastic_internal_variable;
   std::vector<std::vector<dealii::SymmetricTensor<2, dim>>> tmp_stress;
   std::vector<std::vector<dealii::SymmetricTensor<2, dim>>> tmp_back_stress;
@@ -194,8 +223,6 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
   // already reserve the memory.
   unsigned int const n_dofs_per_cell = _fe_collection.max_dofs_per_cell();
   unsigned int const n_old_active_cells = _plastic_internal_variable.size();
-  std::vector<dealii::types::global_dof_index> global_dof_indices(
-      n_dofs_per_cell);
   tmp_plastic_internal_variable.reserve(n_old_active_cells);
   tmp_stress.reserve(n_old_active_cells);
   tmp_back_stress.reserve(_back_stress.size());
@@ -205,7 +232,12 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
     _old_displacement.update_ghost_values();
 
     std::vector<double> cell_values(n_dofs_per_cell);
-    saved_old_displacement.reserve(n_old_active_cells);
+    _saved_old_displacement.resize(n_old_active_cells);
+
+    std::vector<dealii::types::global_dof_index> global_dof_indices(
+        n_dofs_per_cell);
+
+    cell_id = 0;
     for (auto const &cell :
          dealii::filter_iterators(_dof_handler.active_cell_iterators(),
                                   dealii::IteratorFilters::LocallyOwnedCell()))
@@ -226,11 +258,13 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
         // is ignored.
         cell_values.assign(n_dofs_per_cell, 0.);
       }
[…]
