### 2024-09-20 adamantine-sim/adamantine#319 @ericneiva
joss-paper: Minor improvements of software paper
First of all, I would like to commend you for a very clear and well-structured software paper.

Please, find below some minor suggestions for improvement (some of them have already been pointed out by other reviewers):

* [x] _Line 7_ Please cite Trilinos as suggested by the project authors. See [Citing Trilinos](https://trilinos.github.io/cite.html).
* [x] _Line 9_ Direct -> Directed
* [x] _Line 11_ For completeness, I suggest to also mention the other simulated physics (mechanics and thermomechanics).
* [x] _Line 23_ obtained -> obtain
* [x] _Line 31_ Duplicate definition of IR acronym (see _Line 16_)
* [x] _Line 62_ fosr -> for
* [x] _Line 63_ assosciated -> associated
* [x] _Line 70_ stocastic -> stochastic
* [x] _Line 71_ Duplicate definition of EnKF acronym
* [x] _Line 140_ Review DOI
* [x] _Line 145_ Review DOI

See https://github.com/openjournals/joss-reviews/issues/7017
