### 2024-09-20 adamantine-sim/adamantine#320 @ericneiva
joss-paper: Improvements in README.md
Please, find below some suggestions to improve the documentation of the README.md file:

Also note that some of the changes proposed here should be mirrored in the documentation website.

* [x] _Adamantine logo:_ It cannot be seen in dark mode
* [x] _Statement of need:_ I suggest to elaborate a little more at the beginning about what problems the software can solve and who the target audience is, as in the web documentation.

* [x] [p4est](https://www.p4est.org/) should be in small caps and added as a dependency in the bullet list (via deal.ii)
* [x] Perhaps it's worth adding the optional dependencies in the bullet list for clarity and completeness, that is,
  * _Optional_ SEACAS
  * _Optional_ Caliper

#### Docker

* [x] Please, extend the information to match the one of the web documentation.
* [x] Maybe, for the non-experts in docker, indicate how to copy the output results from the container to the host.

#### Input file

* [x] Add or link to an example _input file_ for the purpose of illustration
* [ ] For the sake of clarity, maybe it will help to format the options in a table (name, values, units, required, comments)
* [ ] The input file description is very informative and exhaustive, but as it is so detailed, maybe the README.md will be more readable, if you put it into a separate markdown file and then link to it
* [x] Several parameters lack the units. Here, a non-exhaustive list:
  * [x] *material_height*
  * [x] *deposition_length*
  * [x] *deposition_width*
  * [x] *deposition_lead_time*
  * [x] *deposition_time*
  * [x] *initial_temperature*
  * [x] *new_material_temperature*
  * [x] ...

### Scan path

  * [x] Link to examples of the _segment_ and _event_ formats

### Material deposition

  * [x] Remove `\` from `deposition\_length` and the rest of parameters.
  * [x] Please link to an example file for file-based deposition.

### Example usage

  * [x] Please, link to the examples in the web documentation to avoid searching for them
  * [x] _In the web documentation_ To facilitate running the examples, provide the execution command for each sub-example of the Demo316 and HourGlass folders

See https://github.com/openjournals/joss-reviews/issues/7017
