### 2024-09-20 #7017 @ericneiva
Dear @Rombur, I would like to commend you and your co-authors for the impressive work and the very well-written software paper.

Please find some suggestions for improvement in the issues https://github.com/adamantine-sim/adamantine/issues/319 and https://github.com/adamantine-sim/adamantine/issues/320

To finish my review, I sill need to run the `Demo316`  and `HourGlass` examples from the docker file. I will do this as soon as I can and let you know.
