### 2024-09-20 0edf38e
Use set_active_fe_indices passing
