### 2024-09-20 #7017 @Rombur
Thanks for the comments. Please do  `docker pull rombur/adamantine:1.0` to get the latest version of the Docker image before trying the examples. I have updated the `HourGlass` example to better show how the code scales.
