### 2024-09-20 87c2225
Add command line to run the examples
