### 2024-09-20 #7017 @ericneiva
Dear @Rombur, I was able to run the `Demo316` and `HourGlass` examples and satisfactorily get the same output and observe some scaling. I have updated https://github.com/adamantine-sim/adamantine/issues/320 with a last suggestion.

After resolving and closing the above issues, all will be good from my side.
