### 2024-09-23 5a41656
Add missing units in input file description
