+++ joss/paper.bib
@@ -33,9 +33,10 @@
 }
 
 @Manual{trilinos-website,
-  title = {The {T}rilinos {P}roject {W}ebsite, https://trilinos.github.io},
-  author = {The {T}rilinos {P}roject {T}eam}},
+  title = {The {T}rilinos {P}roject {W}ebsite},
+  author = {The {T}rilinos {P}roject {T}eam},
   year = {2020 (accessed May 22, 2020)},
+  note = {https://trilinos.github.io}
 }
 
 @Article{kokkos2022,
@@ -52,7 +53,7 @@
 @Book{da2016,
   author = {Asch, Mark and Bocquet, Marc and Nodet, Maëlle},
   title = {Data Assimilation},
-  publisher = {Society for Industrial and Applied Mathematics},
+  publisher = {{Society for Industrial and Applied Mathematics}},
   year = {2016},
   doi = {10.1137/1.9781611974546},
   address = {Philadelphia, PA},
@@ -107,11 +108,20 @@
   address = {United States}
 }
 
+@manual{ansys,
+  title        = {ANSYS Documentation},
+  author       = {{ANSYS Inc.}},
+  organization = {ANSYS Inc.},
+  address      = {Canonsburg, PA, USA},
+  year         = {2024},
+  note         = {Release 2024 R1},
+  url          = {https://www.ansys.com/products/structures/ansys-mechanical}
+}
 
 @article{KNAPP2023111904,
 	abstract = {Melt pool scale numerical modeling of additive manufacturing (AM) processes can provide predictive capabilities and theoretical insight into the process-property-structure-performance relationships for AM parts. Despite capabilities of numerical models to solve complex multi-physics problems, it is often important to consider a tradeoff between detailed physics and computational cost. Therefore, sources of uncertainty in both experimental conditions and the parameters needed for modeling require models to be validated against empirical evidence. Here, a method is proposed to calibrate uncertain parameters used in continuum-scale melt pool models for powder bed fusion (PBF) AM. Both a simplified heat transfer model and a heat transfer and fluid flow model were investigated. A surrogate model and Markov chain-based optimization algorithm calibrated melt pool geometry for models within experimental variation of the target melt pool width and depth from the NIST AM-Bench 2018-02 dataset. The melt pool temperature distributions, solidification parameters, and simulated multi-layer solidification microstructures were compared between the two models. Similar results from both models indicate that calibrated, lower fidelity numerical models may be used in place of higher fidelity models to generate melt pool solidification data. These calibrated models therefore enable lower computational cost melt pool simulations without a noticeable decrease in simulation accuracy for grain-scale microstructure simulations.},
 	author = {G.L. Knapp and J. Coleman and M. Rolchigo and M. Stoyanov and A. Plotkowski},
-	doi = {https://doi.org/10.1016/j.commatsci.2022.111904},
+	doi = {10.1016/j.commatsci.2022.111904},
 	issn = {0927-0256},
 	journal = {Computational Materials Science},
 	keywords = {Additive manufacturing, Validation, Calibration, Modeling, Surrogate, Powder bed fusion},
@@ -165,7 +175,7 @@
 @article{KELLER2017244,
[…]
