+++ joss/paper.md
@@ -22,9 +22,11 @@ bibliography: paper.bib
 # Summary
 `Adamantine` is a thermomechanical simulation code that is written in C++ and built on top of deal.II [@dealII95], ArborX [@arborx2020], Trilinos [@trilinos-website], and Kokkos [@kokkos2022]. 
 `Adamantine` was developed with additive manufacturing in mind and it is
-particularly well adapted to simulate fused filament fabrication, direct energy 
-deposition, and powder bed fusion. In order to support these different additive manufacturing processes, 
-`adamantine` solves a nonlinear anisotropic heat equation using the finite element method with adaptive mesh refinement. It can handle materials in
+particularly well adapted to simulate fused filament fabrication, directed energy 
+deposition, and powder bed fusion. `Adamantine` employs the finite element
+method with adaptive mesh refinement to solve a nonlinear anisotropic heat equation, enabling
+support for various additive manufacturing processes. It can also perform
+elastoplastic and thermoelastoplastic simulations. It can handle materials in
 three distinct phases (solid, liquid, and powder) to accurately reflect the
 physical state during different stages of the manufacturing process. To enhance
 simulation accuracy, `adamantine` incorporates data assimilation techniques [@da2016].
@@ -38,7 +40,7 @@ Manufacturing "born-qualified" components, i.e., parts ready for critical
 applications straight from the printer, requires a new approach to additive
 manufacturing (AM). This vision demands not only precise simulations for 
 planning the build but also real-time adjustments throughout the process 
-to obtained the desired thermomechanical evolution of the part. Currently, 
+to obtain the desired thermomechanical evolution of the part. Currently, 
 setting AM process parameters is an expert-driven, often trial-and-error 
 process. Material changes and geometry complexities can lead to unpredictable 
 adjustments in parameters, making a purely empirical approach slow and expensive. We can overcome 
@@ -48,7 +50,7 @@ this by using advanced simulations for both planning and adaptive control.
 parameter planning and adjustment in AM. During the
 planning phase, its capabilities can be leveraged to predict the
 thermomechanical state and optimize process parameters for the desired outcome. 
-For adaptive control, `adamantine` utilizes data from infrared (IR) cameras and 
+For adaptive control, `adamantine` utilizes data from IR cameras and 
 thermocouples. This data is integrated using the Ensemble Kalman Filter (EnKF) method,
 allowing the simulation to constantly adapt and reflect the actual build process.
 
@@ -58,7 +60,7 @@ mid-print, if needed, to ensure that printed parts conform to the necessary tole
 
 While other open-source software like AdditiveFOAM [@additivefoam] excels at heat
 and mass transfer simulations in additive manufacturing, and commercial options
-like Abaqus [@abaqus] and Ansys offer comprehensive thermomechanical capabilities,
+like Abaqus [@abaqus] and Ansys[@ansys] offer comprehensive thermomechanical capabilities,
 `adamantine` stands out for its unique ability to incorporate real-world data
 through data assimilation. This feature allows for potentially more accurate
 simulations, leading to better process optimization and final part quality.
@@ -74,16 +76,16 @@ the isotropic and kinematic hardening described in @borja2013. This allows us to
 model both the change in yield stress and the Bauschinger effect.
 
 ## Thermomechanical simulation
-Thermomechanical simulations in `adamantine` are performed with one-way coupling from the temperature evolution to the mechanical evolution. We neglect the effect of deformation on the thermal simulation. An extra term in the mechanical simulation accounts fosr the eigenstrain assosciated with by thermal expansion of the material [@fung2001; @Megahed2016].
[…]
