### 2024-09-24 adamantine-sim/adamantine#319 @Rombur
Everything on the list is done. @ericneiva I'll let you close the issue if you are happy with the result.
