### 2024-09-24 adamantine-sim/adamantine#319 @ericneiva
Nice! Thanks a lot @Rombur for taking care of all the comments!
