### 2024-09-24 #7017 @Rombur
I think I've done everything asked by @lukashof and @mohd-afeef-badri except for this comment by @mohd-afeef-badri

> Including a brief section that outlines the types of meshes supported by Adamantine and what users can expect during post-processing would significantly enhance the paper's utility for potential users. Perhaps and example simulations that demonstrates this.

What do you mean by "types of meshes"? Do you mean the format of the meshes or the types of elements or both? I am also unsure what are you expectation for "what users can expect during post-processing". I did add example simulations on the website. Does that satisfy your comment?

I have created a new Docker image `rombur/adamantine:1.0-A100` that has GPU support for A100.
