+++ README.md
@@ -1,7 +1,15 @@
-![image](doc/logo/adamantine_wordmark.png)
+![image](doc/logo/adamantine_wordmark_with_bg.png)
 
-`adamantine` stands for *AD*ditive (*A*) *MAN*unifac*T*uring s*I*mulator (*NE*).
-It is an open-source sofware to simulate heat transfer for additive manufacturing.
+`adamantine` is a thermomechanical code for additive manufacturing. It is based
+on [deal.II](https://www.dealii.org),
+[ArborX](https://github.com/arborx/ArborX),
+[Trilinos](https://www.trilinos.github.io), and [Kokkos](https://kokkos.org).
+`adamantine` can simulate the thermomechanical evolution an object undergoes
+during the manufacturing process. It can handle materials in three distinct
+phases (solid, liquid, and powder) to accurately reflect the physical state
+during manufacturing. Experimental data can be used to improve the simulation
+through the use of [Ensemble Kalman
+filter](https://en.wikipedia.org/wiki/Ensemble_Kalman_filter).
 
 ## Installation
 Installing `adamantine` requires:
@@ -10,11 +18,10 @@ Installing `adamantine` requires:
 * CMake: 3.15 or later
 * Boost: 1.70.0 or later
 * ArborX: 1.4.1 or later
-* Trilinos: 14.4.0 or later
-* deal.II: 9.5 or later
-
-You need to compile deal.II with MPI, P4EST, and ArborX. If you want to use Exodus file, you also need Trilinos with SEACAS support.
-`adamantine` also optionally supports profiling through [Caliper](https://github.com/llnl/Caliper).
+* Trilinos: 14.4.0 or later. If you want to *adamantine* to read an Exodus mesh,
+  you need to enable SEACAS support.
+* deal.II: 9.5 or later. You need to compile deal.II with MPI, p4est, and ArborX support.
+* Caliper: 2.10 or later. Optional dependency to profile `adamantine`.
 
 An example on how to install all the dependencies can be found in
 `ci/Dockerfile`.
@@ -47,6 +54,19 @@ using
 ```bash
 docker pull rombur/adamantine:latest
 ```
+The 1.0 release version is available using
+```bash
+docker pull rombur/adamantine:1.0
+```
+To start an interactive container use
+```
+docker run --rm -it rombur/adamantine:1.0 bash
+```
+You will find `adamantine` in `/home/adamantine/bin`. You can mount a folder
+into a Docker image using
+```
+docker run --rm -it -v /path/to/compute/folder:/path/to/image/folder rombur/adamantine:1.0 bash
+```
 
 ## Run
 After compiling `adamantine`, you can run a simulation using
@@ -64,8 +84,11 @@ export DEAL_II_NUM_THREADS=1
 If you use our Docker image, the variable is already set.
 
 ### Input file
-Adamantine supports Boost INFO format and json. The input file is assumed to use
-the INFO format unless the file extension is `.json`.
+`adamantine` supports Boost INFO format and json. The input file is assumed to use
+the INFO format unless the file extension is `.json`. An example of an input
+file can be found
+[here](https://github.com/adamantine-sim/adamantine/blob/master/tests/data/HourGlass_AOP.info).
+
 The following options are available:
 * boundary (required):
   * type: type of boundary: adiabatic, radiative, or convective. Multiple types
@@ -85,7 +108,7 @@ The following options are available:
 * geometry (required):
   * dim: the dimension of the problem (2 or 3, required)
   * material\_height: below this height the domain contains material. Above this
-  height the domain is empty (default value: 1e9)
+  height the domain is empty. The height is in meters (default value: 1e9)
   * use\_powder: the additive manufacturing process use powder: true or false
   (default value: false)
   * if use\_powder is true:
[…]
