### 2024-09-24 adamantine-sim/adamantine#320 @Rombur
@ericneiva I've done all the changes but two for the `input` section. We agree that the `input` section in the README can be improved but we would like more time to find the best way to present the information. If it's ok with you, we will keep the section as-if for now.
