### 2024-09-25 adamantine-sim/adamantine#320 @ericneiva
Hi, @Rombur, sounds good to me. Thanks for applying all the other changes.

I will check off the remaining items in https://github.com/openjournals/joss-reviews/issues/7017 meaning my review is complete.

Do you want to keep this issue open to not forget about the remaining comments or you'd rather close it?
