### 2024-09-25 adamantine-sim/adamantine#320 @Rombur
I would prefer if you close the issue since this is issue is for the paper.
