+++ source/MechanicalPhysics.cc
@@ -76,21 +76,6 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
 void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    setup_dofs(
-        dealii::DoFHandler<dim> const &thermal_dof_handler,
-        dealii::LA::distributed::Vector<double, dealii::MemorySpace::Host> const
-            &temperature,
-        std::vector<bool> const &has_melted,
-        std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
-{
-  _mechanical_operator->update_temperature(thermal_dof_handler, temperature,
-                                           has_melted);
-  setup_dofs(body_forces);
-}
-
-template <int dim, int p_order, typename MaterialStates,
-          typename MemorySpaceType>
-void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
     setup_dofs(std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
 {
   std::cout << "Distribute dofs mechanical" << std::endl;
@@ -156,7 +141,7 @@ unsigned int MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
   }
   return n_local_cells;
 }
-
+/*
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
 void MechanicalPhysics<dim, p_order, MaterialStates,
@@ -203,7 +188,7 @@ void MechanicalPhysics<dim, p_order, MaterialStates,
 
   displacement.update_ghost_values();
   std::swap(_old_displacement, displacement);
-}
+}*/
 
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
@@ -235,15 +220,19 @@ void MechanicalPhysics<dim, p_order, MaterialStates,
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
 void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    prepare_transfer(
-        std::optional<std::reference_wrapper<const dealii::DoFHandler<dim>>>
-            thermal_dof_handler)
+    setup_dofs(
+        dealii::DoFHandler<dim> const &thermal_dof_handler,
+        dealii::LA::distributed::Vector<double, dealii::MemorySpace::Host> const
+            &temperature,
+        std::vector<bool> const &has_melted,
+        std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
 {
-  std::cout << "prepare_transfer" << std::endl;
+  _mechanical_operator->update_temperature(thermal_dof_handler, temperature,
+                                           has_melted);
   // Update the active fe indices, the plastic variables, and the displacement.
   unsigned int const n_quad_pts = _q_collection.max_n_quadrature_points();
   unsigned int cell_id = 0;
-  // std::vector<std::vector<double>> saved_old_displacement;
+  std::vector<std::vector<double>> saved_old_displacement;
   std::vector<std::vector<double>> tmp_plastic_internal_variable;
   std::vector<std::vector<dealii::SymmetricTensor<2, dim>>> tmp_stress;
   std::vector<std::vector<dealii::SymmetricTensor<2, dim>>> tmp_back_stress;
@@ -251,6 +240,8 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
   // already reserve the memory.
   unsigned int const n_dofs_per_cell = _fe_collection.max_dofs_per_cell();
[…]
