### 2024-09-26 #7017 @mohd-afeef-badri
> What do you mean by "types of meshes"? Do you mean the format of the meshes or the types of elements or both? I am also unsure what are you expectation for "what users can expect during post-processing". I did add example simulations on the website. Does that satisfy your comment?

@Rombur thanks for all the efforts concerning the review.
- By "types of meshes" I mean if it should be hexahedral mesh, tetrahedral mesh, polyhedral mesh, ... and what format .msh, .vtk, etc.
- "what users can expect during post-processing", with the example  simulation on website this comment is already taken care of.
