### 2024-09-26 1919df0
Introduce assemble_system
