+++ joss/paper.bib
@@ -14,6 +14,18 @@
   url     = {https://dealii.org/deal95-preprint.pdf}
 }
 
+@ARTICLE{p4est,
+  author = {Carsten Burstedde and Lucas C. Wilcox and Omar Ghattas},
+  title = {{\texttt{p4est}}: Scalable Algorithms for Parallel Adaptive Mesh
+           Refinement on Forests of Octrees},
+  journal = {SIAM Journal on Scientific Computing},
+  volume = {33},
+  number = {3},
+  pages = {1103-1133},
+  year = {2011},
+  doi = {10.1137/100791634}
+}
+
 @Article{arborx2020,
   author = {Lebrun-Grandi\'{e}, D. and Prokopenko, A. and Turcksin, B. and Slattery, S. R.},
   title = {{ArborX}: A Performance Portable Geometric Search Library},
@@ -117,6 +129,75 @@
   note         = {Release 2024 R1},
   url          = {https://www.ansys.com/products/structures/ansys-mechanical}
 }
+
+@book{abaqus,
+  title = "ABAQUS/Standard User's Manual, Version 6.9",
+  author = "Michael Smith",
+  year = "2009",
+  language = "English",
+  publisher = "Dassault Syst{\`e}mes Simulia Corp",
+  address = "United States",
+}
+
+@article{gmsh,
+author = {Geuzaine, Christophe and Remacle, Jean-François},
+title = {Gmsh: A 3-D finite element mesh generator with built-in pre- and post-processing facilities},
+journal = {International Journal for Numerical Methods in Engineering},
+volume = {79},
+number = {11},
+pages = {1309-1331},
+keywords = {computer-aided design, mesh generation, post-processing, finite element method, open-source software},
+doi = {https://doi.org/10.1002/nme.2579},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.2579},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.2579},
+abstract = {Abstract Gmsh is an open-source 3-D finite element grid generator with a build-in CAD engine and post-processor. Its design goal is to provide a fast, light and user-friendly meshing tool with parametric input and advanced visualization capabilities. This paper presents the overall philosophy, the main design choices and some of the original algorithms implemented in Gmsh. Copyright © 2009 John Wiley \& Sons, Ltd.},
+year = {2009}
+}
+
+@manual{comsol,
+  title={COMSOL multiphysics{\textregistered} v6.2},
+  author={{COMSOL AB}},
+  year={2024},
+  url={https://www.comsol.com}
+}
+
+@misc{smesh,
+  title={SMESH Documentation},
+  author={},
+  year={2024},
+  url={https://docs.salome-platform.org/latest/gui/SMESH/index.html}
+}
+
+@misc{exodusii,
+  title = {ExodusII Finite Element Data Model, Version 00},
+  author = {},
+  url = {https://www.osti.gov/biblio/1230926},
+  year = {2005},
+  month = {5}
+}
+
+@manual{tecplot,
+  title        = {Tecplot Documentation},
+  author       = {{Tecplot Inc.}},
+  year         = {2024},
+  url          = {https://tecplot.azureedge.net/products/360/current/360-users-manual.pdf}
+}
+
+@misc{assimp,
+  title = {The Asset-Importer-Lib Documentation},
+  author = {},
+  year = {2024},
+  url = {https://assimp-docs.readthedocs.io/en/latest/}
+}
+
+@Book{vtk,
+  author    = {Will Schroeder and Ken Martin and Bill Lorensen},
+  title     = {The Visualization Toolkit (4th ed.)},
[…]
