### 2024-09-26 017e0a2
try swapping prepare, relevenat dofs
