+++ tests/test_post_processor.cc
@@ -38,7 +38,9 @@ BOOST_AUTO_TEST_CASE(thermal_post_processor)
   geometry_database.put("length_divisions", 4);
   geometry_database.put("height", 6);
   geometry_database.put("height_divisions", 5);
-  adamantine::Geometry<2> geometry(communicator, geometry_database);
+  boost::optional<boost::property_tree::ptree const &> units_optional_database;
+  adamantine::Geometry<2> geometry(communicator, geometry_database,
+                                   units_optional_database);
   // Create the DoFHandler
   dealii::hp::FECollection<2> fe_collection;
   fe_collection.push_back(dealii::FE_Q<2>(2));
@@ -81,8 +83,8 @@ BOOST_AUTO_TEST_CASE(thermal_post_processor)
   beam_database.put("scan_path_file_format", "segment");
   std::vector<std::shared_ptr<adamantine::HeatSource<2>>> heat_sources;
   heat_sources.resize(1);
-  heat_sources[0] =
-      std::make_shared<adamantine::GoldakHeatSource<2>>(beam_database);
+  heat_sources[0] = std::make_shared<adamantine::GoldakHeatSource<2>>(
+      beam_database, units_optional_database);
 
   // Initialize the ThermalOperator
   adamantine::ThermalOperator<2, false, 0, 2, adamantine::SolidLiquidPowder,
@@ -155,7 +157,9 @@ BOOST_AUTO_TEST_CASE(mechanical_post_processor)
   geometry_database.put("height_divisions", 3);
   geometry_database.put("width", 6);
   geometry_database.put("width_divisions", 3);
-  adamantine::Geometry<dim> geometry(communicator, geometry_database);
+  boost::optional<boost::property_tree::ptree const &> units_optional_database;
+  adamantine::Geometry<dim> geometry(communicator, geometry_database,
+                                     units_optional_database);
   // Create the DoFHandler
   dealii::hp::FECollection<dim> fe_collection;
   fe_collection.push_back(dealii::FESystem<dim>(dealii::FE_Q<dim>(2) ^ dim));
