### 2024-09-30 #7017 @lukashof
Hi @lucydot and @Rombur,

sorry for not replying earlier! However, we followed the discussion and ran some of the other examples in the meantime.

Today @FelixFroelich and I found some time to discuss the review.

The addition of the website makes the project more accessible. Thank you for that! It would be great to have some additional information or comparisons on features like the coupled thermomechanical simulations. We understand however, that this might go beyond the scope of the JOSS article and might be the subject of other publications.

Regarding the references of the paper, we think you should change the citation of the ABAQUS user manual. To our knowledge, the version you are citing does not include the AM-plugin, which you probably had in mind, yet. We are also not sure who "Michael Smith" is and why the manual would be attributed to him.

We have checked all points but the one regarding the references and this would conclude the review as far as we are concerned.
