+++ joss/paper.bib
@@ -111,13 +111,14 @@
   url          = {https://doi.org/10.5281/zenodo.8034098}
 }
 
-@book{abaqus,
-  title = {ABAQUS/Standard User's Manual, Version 6.9},
-  author = {Michael Smith},
-  year = {2009},
-  language = {English},
-  publisher = {Dassault Syst{\`e}mes Simulia Corp},
-  address = {United States}
+@manual{abaqus,
+  title        = {Abaqus Documentation},
+  author       = {{Dassault Systèmes Simulia Corp.}},
+  organization = {Dassault Systèmes},
+  address      = {Providence, RI, USA},
+  year         = {2024},
+  note         = {Version 2024},
+  url          = {https://www.3ds.com/products-services/simulia/products/abaqus/}
 }
 
 @manual{ansys,
@@ -128,15 +129,6 @@
   year         = {2024},
   note         = {Release 2024 R1},
   url          = {https://www.ansys.com/products/structures/ansys-mechanical}
-}
-
-@book{abaqus,
-  title = "ABAQUS/Standard User's Manual, Version 6.9",
-  author = "Michael Smith",
-  year = "2009",
-  language = "English",
-  publisher = "Dassault Syst{\`e}mes Simulia Corp",
-  address = "United States",
 }
 
 @article{gmsh,
