### 2024-09-30 #7017 @Rombur
I have updated the ABAQUS reference. It is more in line with the Ansys reference.
