+++ source/MechanicalPhysics.cc
@@ -211,19 +211,17 @@ template <int dim, int p_order, typename MaterialStates,
 void MechanicalPhysics<dim, p_order, MaterialStates,
                        MemorySpaceType>::prepare_transfer_mpi()
 {
-  _solution_transfer.prepare_for_coarsening_and_refinement(_old_displacement);
-  /*
-    const dealii::IndexSet &locally_owned_dofs =
+  //_solution_transfer.prepare_for_coarsening_and_refinement(_old_displacement); 
+   const dealii::IndexSet &locally_owned_dofs =
         _dof_handler.locally_owned_dofs();
     const dealii::IndexSet locally_relevant_dofs =
         dealii::DoFTools::extract_locally_relevant_dofs(_dof_handler);
 
-    dealii::LA::distributed::Vector<double, dealii::MemorySpace::Host>
-        relevant_displacement(locally_owned_dofs, locally_relevant_dofs,
+        _relevant_displacement.reinit(locally_owned_dofs, locally_relevant_dofs,
                               MPI_COMM_WORLD);
-    relevant_displacement = _old_displacement;
+    _relevant_displacement = _old_displacement;
     _solution_transfer.prepare_for_coarsening_and_refinement(
-        relevant_displacement);*/
+        _relevant_displacement);
 }
 
 template <int dim, int p_order, typename MaterialStates,
