### 2024-10-01 #7017 @lucydot
There are a few small edits for the paper:
- ` Ansys(ANSYS Inc., 2024)` needs a space
- The Borja reference in the main text is not formatted correctly
- The (I assume) list under `MPI support` is not formatted correctly
