+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## [1.0](https://github.com/adamantine-sim/adamantine/tree/release/1.0) (2024-10-01)
+
+Initial release.
