### 2024-10-01 895e2a4
Remove first sentence of index
