### 2024-10-01 d8f1639
More ghosted vectors; passing tests
