### 2024-10-01 #7017 @Rombur
I have created the release [1.0](https://github.com/adamantine-sim/adamantine/releases/tag/v1.0) and I have a doi from Zenodo https://doi.org/10.5281/zenodo.13869657
