### 2024-10-01 fcc4099
Trabsfer plasic and stress
