### 2024-10-03 2b04d03
Clean-up MechanicalPhysics
