+++ source/MechanicalPhysics.cc
@@ -48,10 +48,22 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
   // Solve the mechanical problem only on the part of the domain that has solid
   // material.
-
-  std::cout << "Setting active fe_indices mechanical 0" << std::endl;
-
-  unsigned int n_local_cells = set_active_fe_indices();
+  unsigned int n_local_cells = 0;
+  for (auto const &cell :
+       dealii::filter_iterators(_dof_handler.active_cell_iterators(),
+                                dealii::IteratorFilters::LocallyOwnedCell()))
+  {
+    if (_material_properties.get_state_ratio(
+            cell, MaterialStates::State::solid) > 0.99)
+    {
+      cell->set_active_fe_index(0);
+      ++n_local_cells;
+    }
+    else
+    {
+      cell->set_active_fe_index(1);
+    }
+  }
 
   // Create the mechanical operator
   _mechanical_operator = std::make_unique<
@@ -79,19 +91,9 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
 void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    setup_dofs(
-        std::vector<std::shared_ptr<BodyForce<dim>>> const & /*body_forces*/)
+    setup_dofs(std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
 {
-  std::cout << "Distribute dofs mechanical" << std::endl;
-  //_old_displacement.print(std::cout);
-  //_dof_handler.locally_owned_dofs().print(std::cout);
-  // std::cout << "dofs before setup_dofs " << _dof_handler.n_dofs() <<
-  // std::endl;
   _dof_handler.distribute_dofs(_fe_collection);
-  //       std::cout << "dofs after setup_dofs " << _dof_handler.n_dofs() <<
-  //       std::endl;
-  //_old_displacement.print(std::cout);
-  //_dof_handler.locally_owned_dofs().print(std::cout);
   dealii::IndexSet locally_relevant_dofs;
   dealii::DoFTools::extract_locally_relevant_dofs(_dof_handler,
                                                   locally_relevant_dofs);
@@ -107,114 +109,16 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
       _dof_handler, 4, dealii::Functions::ZeroFunction<dim>(dim),
       _affine_constraints);
   _affine_constraints.close();
-}
 
-template <int dim, int p_order, typename MaterialStates,
-          typename MemorySpaceType>
-void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    assemble_system(
-        std::vector<std::shared_ptr<BodyForce<dim>>> const &body_forces)
-{
   _mechanical_operator->reinit(_dof_handler, _affine_constraints, _q_collection,
                                body_forces);
 }
 
 template <int dim, int p_order, typename MaterialStates,
           typename MemorySpaceType>
-unsigned int MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
-    set_active_fe_indices(
-        std::optional<std::reference_wrapper<dealii::DoFHandler<dim> const>>
-            thermal_dof_handler)
-
-{
-  unsigned int n_local_cells = 0;
-  for (auto const &cell :
-       dealii::filter_iterators(_dof_handler.active_cell_iterators(),
-                                dealii::IteratorFilters::LocallyOwnedCell()))
-  {
-    auto current_fe_index = cell->future_fe_index();
[…]
