### 2024-10-03 10d8125
Clean up thermal physics
