+++ tests/test_material_deposition.cc
@@ -331,7 +331,7 @@ BOOST_AUTO_TEST_CASE(material_deposition)
 
       thermal_physics.add_material_start(
           elements_to_activate, deposition_cos, deposition_sin, has_melted,
-          activation_start, activation_end, initial_temperature, solution);
+          activation_start, activation_end, solution);
 
 #ifdef ADAMANTINE_WITH_CALIPER
       CALI_MARK_BEGIN("refine triangulation");
@@ -346,9 +346,7 @@ BOOST_AUTO_TEST_CASE(material_deposition)
       CALI_MARK_END("refine triangulation");
 #endif
 
-      thermal_physics.add_material_end(
-          elements_to_activate, deposition_cos, deposition_sin, has_melted,
-          activation_start, activation_end, initial_temperature, solution);
+      thermal_physics.add_material_end(initial_temperature, solution);
     }
 
     time =
