### 2024-10-03 #7017 @kyleniemeyer
I think the `https://doi.org/` at the beginning should be removed—DOIs need to be specified without that.
