### 2024-10-03 d6040ff
Remove extra https in doi reference
