+++ joss/paper.bib
@@ -139,7 +139,7 @@ volume = {79},
 number = {11},
 pages = {1309-1331},
 keywords = {computer-aided design, mesh generation, post-processing, finite element method, open-source software},
-doi = {https://doi.org/10.1002/nme.2579},
+doi = {10.1002/nme.2579},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nme.2579},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/nme.2579},
 abstract = {Abstract Gmsh is an open-source 3-D finite element grid generator with a build-in CAD engine and post-processor. Its design goal is to provide a fast, light and user-friendly meshing tool with parametric input and advanced visualization capabilities. This paper presents the overall philosophy, the main design choices and some of the original algorithms implemented in Gmsh. Copyright © 2009 John Wiley \& Sons, Ltd.},
