+++ joss/paper.bib
@@ -215,7 +215,7 @@ year = {2009}
 	date-modified = {2024-06-05 08:12:23 -0400},
 	doi = {10.1186/s40192-016-0047-2},
 	id = {Megahed2016},
-	isbn = {2193-9772},
+	issn = {2193-9772},
 	journal = {Integrating Materials and Manufacturing Innovation},
 	number = {1},
 	pages = {61--93},
@@ -234,7 +234,7 @@ year = {2009}
 	date-modified = {2024-06-05 09:20:50 -0400},
 	doi = {10.1007/BF02667333},
 	id = {Goldak1984},
-	isbn = {2379-0229},
+	issn = {2379-0229},
 	journal = {Metallurgical Transactions B},
 	number = {2},
 	pages = {299--305},
