### 2024-10-04 220e04b
Clen up tests/test_mechanical_physics.cc
