### 2024-10-08 e57f505
Run test_integration_thermoelastic in parallel
