+++ source/MechanicalPhysics.cc
@@ -48,7 +48,8 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
   // Solve the mechanical problem only on the part of the domain that has solid
   // material.
-  unsigned int n_local_cells = 0;
+  unsigned int n_active_cells =
+      _dof_handler.get_triangulation().n_active_cells();
   for (auto const &cell :
        dealii::filter_iterators(_dof_handler.active_cell_iterators(),
                                 dealii::IteratorFilters::LocallyOwnedCell()))
@@ -57,7 +58,6 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
             cell, MaterialStates::State::solid) > 0.99)
     {
       cell->set_active_fe_index(0);
-      ++n_local_cells;
     }
     else
     {
@@ -72,19 +72,26 @@ MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
   // Create the data used to compute the stress tensor
   unsigned int const n_quad_pts = _q_collection.max_n_quadrature_points();
-  _plastic_internal_variable.reserve(n_local_cells);
-  for (auto const &cell :
-       dealii::filter_iterators(_dof_handler.active_cell_iterators(),
-                                dealii::IteratorFilters::LocallyOwnedCell()))
+  _plastic_internal_variable.reserve(n_active_cells);
+
+  for (auto const &cell : _dof_handler.active_cell_iterators())
   {
-    auto elastic_limit = _material_properties.get_mechanical_property(
-        cell, StateProperty::elastic_limit);
-    _plastic_internal_variable.emplace_back(
-        std::vector<double>(n_quad_pts, elastic_limit));
+    if (cell->is_locally_owned())
+    {
+      auto elastic_limit = _material_properties.get_mechanical_property(
+          cell, StateProperty::elastic_limit);
+      _plastic_internal_variable.emplace_back(
+          std::vector<double>(n_quad_pts, elastic_limit));
+    }
+    else
+    {
+      _plastic_internal_variable.emplace_back(std::vector<double>(
+          n_quad_pts, std::numeric_limits<double>::signaling_NaN()));
+    }
   }
-  _stress.resize(n_local_cells,
+  _stress.resize(n_active_cells,
                  std::vector<dealii::SymmetricTensor<2, dim>>(n_quad_pts));
-  _back_stress.resize(n_local_cells,
+  _back_stress.resize(n_active_cells,
                       std::vector<dealii::SymmetricTensor<2, dim>>(n_quad_pts));
 }
 
@@ -228,7 +235,6 @@ void MechanicalPhysics<dim, p_order, MaterialStates,
       {
         for (unsigned int i = 0; i < n_doubles_per_quad_stress; ++i)
         {
-          std::cout << _stress[cell_id].size() << std::endl;
           _stress[cell_id][quad].access_raw_entry(i) =
               data_to_unpack[cell_id][stress_offset +
                                       quad * n_doubles_per_quad_stress + i];
@@ -277,79 +283,87 @@ void MechanicalPhysics<dim, p_order, MaterialStates, MemorySpaceType>::
 
     std::vector<double> cell_values(n_dofs_per_cell);
     saved_old_displacement.reserve(n_old_active_cells);
-    for (auto const &cell :
-         dealii::filter_iterators(_dof_handler.active_cell_iterators(),
-                                  dealii::IteratorFilters::LocallyOwnedCell()))
+    for (auto const &cell : _dof_handler.active_cell_iterators())
     {
-      auto fe_index = cell->active_fe_index();
-      if (fe_index == 0)
+      if (cell->is_locally_owned())
       {
[…]
