### 2024-10-14 3afa9fa
Add ClosestSourcePointAdaptation
