### 2024-10-17 #7017 @kyleniemeyer
Hi @Rombur, just doing some final checks before publishing.

In the paper, can you add a space between "p4est(Burstedde et al., 2011)" (before the citation command)? Also, in the Trilinos citation, please wrap the team author in another set of `{}` (right now it is being interpreted and abbreviated as a person's name), and also move the accessed date out of the year field and to a `note` field.
