+++ joss/paper.bib
@@ -46,9 +46,9 @@
 
 @Manual{trilinos-website,
   title = {The {T}rilinos {P}roject {W}ebsite},
-  author = {The {T}rilinos {P}roject {T}eam},
-  year = {2020 (accessed May 22, 2020)},
-  note = {https://trilinos.github.io}
+  author = {{The {T}rilinos {P}roject {T}eam}},
+  year = {2020},
+  note = {https://trilinos.github.io, (accessed May 22, 2020) }
 }
 
 @Article{kokkos2022,
