### 2024-10-17 7e5285f
Avoid using MPI_COMM_WORLD
