### 2024-10-17 af55928
Drop redundant variables
